Successful HW18
Test was successful this time.
