+++ src/test/resources/TestNG.xml
@@ -7,6 +7,7 @@
         <classes>
             <class name="LoginTests"/>
             <class name="BaseTest"/>
+            <class name="Homework18"/>
         </classes>
     </test>
 </suite>
